+++ CHANGES.rst
@@ -35,4 +35,5 @@ v0.1.6
 v0.1.7[unreleased]
 -----------
 * Add codecov badge to README.rst
+* Test async email triggering
 * [TODO] Add better test cases covering UserCreateViewSet and PasswordRecoveryViewSet